# docs/swagger.yaml
basePath: /api
definitions:
...
  ChangePasswordRequest:
    properties:
      confirmNewPassword:
        description: Must be the same as the newPassword value.
        type: string
      currentPassword:
        type: string
      newPassword:
        description: Min 6, max 150 characters.
        type: string
    type: object
...
  EditUser:
    properties:
      email:
        description: Must be unique
        type: string
      image:
        description: image/png or image/jpeg
        format: binary
        type: string
      username:
        description: Min 3, max 30 characters.
        type: string
    type: object
  ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/HttpError'
    type: object
  ErrorsResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/FieldError'
        type: array
    type: object
  FieldError:
    properties:
      field:
        description: The property containing the error
        type: string
      message:
        description: The specific error message
        type: string
    type: object
...
  HttpError:
    properties:
      message:
        description: The specific error message
        type: string
      type:
        description: The Http Response as a string
        type: string
    type: object
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not a json object
        type: boolean
    type: object
  User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      image:
        type: string
      isOnline:
        type: boolean
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:<PORT>
info:
  contact: {}
  description: Xonia REST API . This service uses sessions for authentication
  license:
    name: GNUv3
  title: Xonia API
  version: "2.0.0"
paths:
  /account:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Current User
      tags:
      - Account
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Update Account
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/EditUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update Current User
      tags:
      - Account
...
  /account/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Change Password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Change Current User's Password
      tags:
      - Account
...
swagger: "2.0"